School Models/Repo

+++ src/main/java/com/tiniycompany/school/repositories/VehiculeRepository.java
@@ -3,5 +3,5 @@ package com.tiniycompany.school.repositories;
 import com.tiniycompany.school.models.Vehicule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehiculeRepository extends JpaRepository<Vehicule, Long> {
+public interface VehiculeRepository extends JpaRepository<Vehicule, Integer> {
 }
